requirements.txt
altgraph==0.17.3
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-tailwind==0.5.0
Django==4.1.7
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-jquery==3.1.0
django-paypal==2.0
future==0.18.3
idna==3.4
pefile==2023.2.7
Pillow==9.4.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyperclip==1.8.2
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.7
pywin32-ctypes==0.2.0
requests==2.31.0
sqlparse==0.4.3
tk==0.1.0
tzdata==2022.7
urllib3==2.0.3
